make KaHyPar and extension
