+++ examples/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+OptimalBranching = "96e750e4-2b4b-404e-96bc-6769eb2ae924"
+SCIP = "82193955-e24f-5292-bf16-6f2c5261a85f"
