+++ examples/rule_discovery.jl
@@ -1,6 +1,7 @@
 # # Automatic rule discovery
 using OptimalBranching.OptimalBranchingMIS.Graphs, OptimalBranching
 using OptimalBranching.OptimalBranchingCore, OptimalBranching.OptimalBranchingCore.BitBasis
+using SCIP
 
 # This function generates the tree-like N3 neighborhood of g0.
 function tree_like_N3_neighborhood(g0::SimpleGraph)
@@ -20,7 +21,7 @@ function solve_opt_rule(branching_region, graph, vs)
     ## Use default solver and measure
     m = D3Measure()
     table_solver = TensorNetworkSolver(; prune_by_env=true)
-    set_cover_solver = IPSolver(verbose=true)
+    set_cover_solver = IPSolver(optimizer= SCIP.Optimizer, verbose=true)
 
     ## Pruning irrelevant entries
     ovs = OptimalBranchingMIS.open_vertices(graph, vs)
