+++ lib/OptimalBranchingMIS/src/OptimalBranchingMIS.jl
@@ -7,11 +7,10 @@ using Combinatorics
 using EliminateGraphs, EliminateGraphs.Graphs
 using GenericTensorNetworks
 using SparseArrays
-using KaHyPar
 
 export MISProblem
 export MISReducer, XiaoReducer, TensorNetworkReducer
-export MinBoundarySelector, MinBoundaryHighDegreeSelector,KaHyParSelector
+export MinBoundarySelector, MinBoundaryHighDegreeSelector, KaHyParSelector
 export TensorNetworkSolver
 export NumOfVertices, D3Measure
