+++ lib/OptimalBranchingMIS/test/runtests.jl
@@ -1,5 +1,4 @@
 using OptimalBranchingMIS
-using TropicalGEMM
 using Test
 
 @testset "branch" begin
